Move filter dataclasses to common
